@@ -1,42 +1,102 @@
 {
   "name": "autostaker-extension",
   "version": "1.0.0",
-  "description": "",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack",
-    "build-full": "fullBuild=true webpack",
-    "watch": "webpack --watch"
+  "description": "Professional Chrome extension for automated yield farming on Terra's Mirror Protocol",
+  "author": "Autostaker Team",
+  "license": "MIT",
+  "private": true,
+  "engines": {
+    "node": ">=16.0.0",
+    "npm": ">=7.0.0"
   },
-  "keywords": [],
-  "license": "ISC",
-  "devDependencies": {
-    "@babel/core": "^7.12.3",
-    "@babel/preset-env": "^7.12.1",
-    "@babel/preset-react": "^7.12.1",
-    "@hot-loader/react-dom": "^17.0.0-rc.2",
-    "@types/chrome": "0.0.125",
-    "@types/react": "^16.9.53",
-    "@types/react-dom": "^16.9.8",
-    "babel-loader": "^8.1.0",
-    "copy-webpack-plugin": "^6.2.1",
-    "css-loader": "^5.0.0",
-    "file-loader": "^6.1.1",
-    "style-loader": "^2.0.0",
-    "ts-loader": "^8.0.5",
-    "typescript": "^4.0.3",
-    "url-loader": "^4.1.1",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^4.0.0",
-    "webpack-dev-server": "^3.11.0"
+  "scripts": {
+    "dev": "NODE_ENV=development webpack --watch --mode development",
+    "build": "NODE_ENV=production webpack --mode production",
+    "build:dev": "NODE_ENV=development webpack --mode development",
+    "build:prod": "npm run clean && npm run build",
+    "clean": "rm -rf extension/*.js extension/*.map",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "test:coverage": "jest --coverage",
+    "lint": "eslint src --ext .ts,.tsx",
+    "lint:fix": "eslint src --ext .ts,.tsx --fix",
+    "format": "prettier --write \"src/**/*.{ts,tsx,js,jsx,json,css,md}\"",
+    "format:check": "prettier --check \"src/**/*.{ts,tsx,js,jsx,json,css,md}\"",
+    "type-check": "tsc --noEmit",
+    "validate": "npm run type-check && npm run lint && npm run format:check && npm run test",
+    "prepare": "husky install",
+    "pre-commit": "lint-staged",
+    "analyze": "webpack-bundle-analyzer extension/stats.json"
   },
   "dependencies": {
-    "@mirror-protocol/mirror.js": "^2.1.0",
-    "@terra-money/terra.js": "^2.0.9",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-hot-loader": "^4.13.0",
-    "react-stl-viewer": "^1.0.3",
-    "stl-viewer": "^0.12.0"
-  }
+    "@terra-money/terra.js": "^3.1.9",
+    "@mirror-protocol/mirror.js": "^2.2.4",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.22.0",
+    "@babel/preset-env": "^7.22.0",
+    "@babel/preset-react": "^7.22.0",
+    "@babel/preset-typescript": "^7.22.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^12.1.5",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/chrome": "^0.0.237",
+    "@types/jest": "^29.5.0",
+    "@types/node": "^20.2.0",
+    "@types/react": "^17.0.60",
+    "@types/react-dom": "^17.0.20",
+    "@typescript-eslint/eslint-plugin": "^5.59.0",
+    "@typescript-eslint/parser": "^5.59.0",
+    "babel-loader": "^9.1.2",
+    "copy-webpack-plugin": "^11.0.0",
+    "css-loader": "^6.8.0",
+    "eslint": "^8.40.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "html-webpack-plugin": "^5.5.1",
+    "husky": "^8.0.3",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^29.5.0",
+    "jest-environment-jsdom": "^29.5.0",
+    "lint-staged": "^13.2.2",
+    "prettier": "^2.8.8",
+    "style-loader": "^3.3.3",
+    "ts-jest": "^29.1.0",
+    "typescript": "^5.0.4",
+    "webpack": "^5.85.0",
+    "webpack-bundle-analyzer": "^4.8.0",
+    "webpack-cli": "^5.1.0"
+  },
+  "lint-staged": {
+    "*.{ts,tsx}": [
+      "eslint --fix",
+      "prettier --write"
+    ],
+    "*.{js,jsx,json,css,md}": [
+      "prettier --write"
+    ]
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/your-org/autostaker-ext.git"
+  },
+  "bugs": {
+    "url": "https://github.com/your-org/autostaker-ext/issues"
+  },
+  "homepage": "https://github.com/your-org/autostaker-ext#readme",
+  "keywords": [
+    "chrome-extension",
+    "terra",
+    "mirror-protocol",
+    "defi",
+    "automation",
+    "staking",
+    "typescript",
+    "react"
+  ]
 }
